// galaxyui/src/app/resources/provider-namespaces/provider-namespace.ts
export class ProviderNamespace {
    id: number;
    name: string;
    description: string;
    display_name: string;
    avatar_url: string;
    location: string;
    company: string;
    email: string;
    html_url: string;
    followers: number;
    provider: number;
    provider_name: string;
    related: object;
    summary_fields: any;
}